requirements.txt
biopython==1.81
streamlit==1.28.1